+++ library/src/main/java/org/mustangproject/Allowance.java
@@ -39,9 +39,8 @@ public class Allowance extends Charge {
 			return totalAmount;
 		} else if (percent!=null) {
 			BigDecimal singlePrice=currentItem.getValue().multiply(BigDecimal.ONE.subtract(getPercent().divide(new BigDecimal(100))));
-//			BigDecimal singlePrice=currentItem.getValue().multiply(BigDecimal.ONE.subtract(getPercent().divide(new BigDecimal(100))));
 			BigDecimal singlePriceDiff=currentItem.getValue().subtract(singlePrice);
-			return singlePriceDiff;
+			return singlePriceDiff.multiply(currentItem.getQuantity());
 		} else {
 			throw new RuntimeException("percent must be set");
 		}