+++ library/src/main/java/org/mustangproject/Item.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import org.mustangproject.ZUGFeRD.IReferencedDocument;
 import org.mustangproject.ZUGFeRD.IZUGFeRDAllowanceCharge;
 import org.mustangproject.ZUGFeRD.IZUGFeRDExportableItem;
+import org.mustangproject.ZUGFeRD.LineCalculator;
 import org.mustangproject.util.NodeMap;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
@@ -203,9 +204,6 @@ public class Item implements IZUGFeRDExportableItem {
 					}
 					if (amountString != null) {
 						izac.setTotalAmount(new BigDecimal(amountString));
-						if (percentString != null && (!percentString.equals("0"))) {
-							izac.setTotalAmount(new BigDecimal(amountString).divide(getQuantity()));
-						}
 					}
 					if (basisAmountString != null) {
 						izac.setBasisAmount(new BigDecimal(basisAmountString));