+++ library/src/main/java/org/mustangproject/ZUGFeRD/TransactionCalculator.java
@@ -195,7 +195,7 @@ public class TransactionCalculator implements IAbsoluteValueProvider {
 				percent = currentItem.getProduct().getVATPercent();
 			}
 			if (percent != null) {
-				LineCalculator lc = new LineCalculator(currentItem);
+				LineCalculator lc = currentItem.getCalculation();
 				VATAmount itemVATAmount = new VATAmount(lc.getItemTotalNetAmount(), lc.getItemTotalVATAmount(),
 					currentItem.getProduct().getTaxCategoryCode(), vatDueDateTypeCode);
 				String reasonText = currentItem.getProduct().getTaxExemptionReason();
@@ -265,7 +265,7 @@ public class TransactionCalculator implements IAbsoluteValueProvider {
 			}
 			if (percent != null)
 			{
-				final LineCalculator lc = new LineCalculator(currentItem);
+				final LineCalculator lc = currentItem.getCalculation();
 				final VATAmount itemVATAmount = new VATAmount(lc.getItemTotalNetAmount(), lc.getItemTotalVATAmount(),
 					currentItem.getProduct().getTaxCategoryCode(), vatDueDateTypeCode, percent);
 				final String reasonText = currentItem.getProduct().getTaxExemptionReason();