+++ validator/src/main/java/org/mustangproject/validator/XMLValidator.java
@@ -374,6 +374,14 @@ public class XMLValidator extends Validator {
 							addUnsupportedProfileResultItem();
 						}
 					}
+				} else {
+					// no CII -> has to be UBL
+					if (context.hasPDF()) {
+						final ValidationResultItem vri = new ValidationResultItem(ESeverity.error, "Factur-X/ZUGFeRD and Order-X are always strictly CII only, no UBL allowed.").setSection(17)
+							.setPart(EPart.fx);
+						context.addResultItem(vri);
+
+					}
 				}
 
 				if (xsltFilename != null) {
@@ -536,12 +544,15 @@ public class XMLValidator extends Validator {
 						}
 
 						ESeverity severity;
+						Node failNode = currentFailNode.getAttributes().getNamedItem("flag");
+						String failVal = failNode == null ? null : failNode.getNodeValue();
 						if (defaultSeverity == ESeverity.notice) {
 							severity = defaultSeverity;
-						} else if (currentFailNode.getAttributes().getNamedItem("flag") != null
-							&& "warning".equals(currentFailNode.getAttributes().getNamedItem("flag").getNodeValue())) {
+						} else if ("warning".equals(failVal)) {
 							// the XR issues warnings with flag=warning
 							severity = ESeverity.warning;
+						} else if ("information".equals(failVal)) {
+							severity = ESeverity.notice;
 						} else {
 							severity = ESeverity.error;
 						}